Add AssertingExceptionsTests

+++ 01-TestesDeUnidade/Demo/Calculadora.cs
@@ -7,7 +7,7 @@
             return v1 + v2;
         }
 
-        public double Dividir(double v1, double v2)
+        public int Dividir(int v1, int v2)
         {
             return v1 / v2;
         }
